backend/backend_backup/backend/backend_backup/requirements.txt
# Core Requirements
flask==2.3.2
gunicorn==20.1.0

# Use Pillow 10+ for Python 3.13 compatibility
pillow==10.1.0  # Updated version
numpy==1.26.4   # Updated for compatibility

# Only include opencv if absolutely needed
# opencv-python-headless==4.9.0.80  # Uncomment if required

# API Support
flask-cors==4.0.0
python-dotenv==1.0.0
whitenoise==6.6.0# Core Flask
flask==2.3.2
gunicorn==20.1.0

# Database (if needed)
flask-sqlalchemy==3.0.3
psycopg2-binary==2.9.6

# Image Processing
opencv-python-headless==4.7.0.72
pillow==9.5.0
numpy==1.24.3

# API & Security
flask-cors==3.0.10
python-dotenv==1.0.0

# Production
whitenoise==6.4.0
